Move private #includes into separate directory.
The installation glob will not match these files, so they will remain
private.
